pyproject.toml
[project]
name = "habit_tracker"
version = "0.1.0"
description = "habit tracker api"
authors = [{name = "Justin Jones", email = "kugelblitz104@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv",
    "sqlalchemy",
    "fastapi[standard]",
    "pydantic[email]",
    "alembic>=1.17.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.11.0",
    "asyncpg>=0.30.0",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
]

[tool.uv]
package = true
native-tls = true

[dependency-groups]
test = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "factory-boy>=3.3.3",
]

[project.scripts]
habit-tracker = "habit_tracker.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "asyncio: mark test as an async test",
]
